backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
cryptography==41.0.7
